+++ smb/server/filevfs/filevfs.go
@@ -42,8 +42,8 @@ import (
 	"time"
 
 	"github.com/jfjallid/go-smb/smb"
-	"github.com/jfjallid/go-smb/smb/encoder"
 	"github.com/jfjallid/go-smb/smb/server"
+	"github.com/jfjallid/go-smb/smb/unicode"
 	"github.com/jfjallid/golog"
 )
 
@@ -702,7 +702,7 @@ func (fs *FS) SetFileInfo(ctx context.Context, h server.Handle, infoClass byte, 
 		if int(nameLen)+20 > len(raw) {
 			return smb.StatusInfoLengthMismatch, nil
 		}
-		newDisplay, err := encoder.FromUnicodeString(raw[20 : 20+nameLen])
+		newDisplay, err := unicode.FromUnicodeString(raw[20 : 20+nameLen])
 		if err != nil {
 			return smb.StatusInvalidParameter, nil
 		}